refactor(dashboard): dedupe Agent Chat setup UI fixes NV-8553
Extract useAgentChatPrompt and shared setup steps, inline connected recap into the integration guide, and fix stale Connected semantics comments.
